+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotFile.java
@@ -2,6 +2,7 @@ package dev.xeonkryptos.xeonrobotframeworkplugin.psi.element;
 
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiFile;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.dto.ImportType;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
@@ -19,7 +20,7 @@ public interface RobotFile extends PsiFile {
      *
      * @return a collection of keyword files that this files knows about.
      */
-    @NotNull Collection<KeywordFile> collectImportedFiles(boolean includeTransitive);
+    @NotNull Collection<KeywordFile> collectImportedFiles(boolean includeTransitive, ImportType... importTypes);
 
     @NotNull Collection<VirtualFile> findImportedFilesWithLibraryName(@NotNull String libraryName);